adding flyway

+++ src/main/java/org/example/tnal_youth_backend/dashboard/dto/MetricResponse.java
@@ -3,7 +3,7 @@ package org.example.tnal_youth_backend.dashboard.dto;
 import java.math.BigDecimal;
 
 public record MetricResponse(
-        BigDecimal value,
+        Long value,
         BigDecimal changePercent
 ) {
 }
